Reorganised ESP32 firmware
